requirements.txt
fastapi==0.78.0
SQLAlchemy==1.4.37
oauthlib==3.2.0
pycrypto==2.6.1
PyJWT==1.7.1
bcrypt==3.2.2
cryptography==37.0.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.9.1
python-jose==3.3.0
slowapi==0.1.5
urllib3==1.26.9
uvicorn==0.17.6

